# specs/002-nfse-query/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NFS-e Query API
  description: |
    API for querying and retrieving NFS-e (Nota Fiscal de Serviço Eletrônica) documents
    from the Sistema Nacional NFS-e.

    This API extends the NFS-e Emission API (001-nfse-emission-core) with read operations.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://api.staging.example.com/v1
    description: Staging
security:
  - ApiKeyAuth: []
...
paths:
  /nfse/{chaveAcesso}:
    get:
      tags:
        - NFS-e Query
      summary: Query NFS-e by access key
      description: |
        Retrieves a complete NFS-e document using its 50-character access key (chaveAcesso).
        Returns parsed JSON data and the complete signed XML.
      operationId: queryNFSe
      parameters:
        - name: chaveAcesso
          in: path
          required: true
          description: 50-character NFS-e access key
          schema:
            type: string
            pattern: '^NFSe[A-Za-z0-9]{46}$'
            minLength: 50
            maxLength: 50
          example: "NFSe3550308202601081123456789012300000000000012310"
      responses:
        '200':
          description: NFS-e found and returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NFSeQueryResponse'
        '400':
          description: Invalid access key format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "INVALID_ACCESS_KEY"
                  message: "Access key must be exactly 50 alphanumeric characters starting with 'NFSe'"
        '404':
          description: NFS-e not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "NFSE_NOT_FOUND"
                  message: "No NFS-e found for the given access key"
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until rate limit resets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Government API unavailable
          headers:
            Retry-After:
              schema:
                type: integer
              description: Suggested retry delay in seconds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    NFSeQueryResponse:
      type: object
      required:
        - chave_acesso
        - numero
        - data_emissao
        - status
        - prestador
        - servico
        - valores
        - xml
      properties:
        chave_acesso:
          type: string
          description: 50-character access key
          example: "NFSe3550308202601081123456789012300000000000012310"
        numero:
          type: string
          description: NFS-e number
          example: "000000123"
        data_emissao:
          type: string
          format: date-time
          description: Emission date/time (ISO 8601)
          example: "2026-01-08T10:30:00-03:00"
        status:
          type: string
          description: NFS-e status code
          example: "100"
        prestador:
          type: object
          required:
            - documento
            - nome
            - municipio
          properties:
            documento:
              type: string
              description: CNPJ or CPF
              example: "12345678000199"
            nome:
              type: string
              description: Provider name
              example: "Empresa Exemplo LTDA"
            municipio:
              type: string
              description: Municipality name
              example: "São Paulo"
        tomador:
          type: object
          nullable: true
          properties:
            documento:
              type: string
              nullable: true
              description: CNPJ or CPF (optional for consumer sales)
            nome:
              type: string
              description: Taker name
        servico:
          type: object
          required:
            - codigo_nacional
            - descricao
            - local_prestacao
          properties:
            codigo_nacional:
              type: string
              description: National tax code (cTribNac)
              example: "010101"
            descricao:
              type: string
              description: Service description
            local_prestacao:
              type: string
              description: Service location
        valores:
          type: object
          required:
            - valor_servico
            - base_calculo
            - valor_liquido
          properties:
            valor_servico:
              type: number
              format: double
              description: Service value
              example: 1000.00
            base_calculo:
              type: number
              format: double
              description: Tax base
              example: 1000.00
            aliquota:
              type: number
              format: double
              nullable: true
              description: ISSQN rate (%)
              example: 5.00
            valor_issqn:
              type: number
              format: double
              nullable: true
              description: ISSQN value
              example: 50.00
            valor_liquido:
              type: number
              format: double
              description: Net value
              example: 950.00
        xml:
          type: string
          description: Complete signed NFS-e XML
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "NFSE_NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "No NFS-e found for the given access key"
            details:
              type: string
              nullable: true
              description: Additional error context
            field:
              type: string
              nullable: true
              description: Field name for validation errors